requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
pytz
wikipedia
geopy
timezonefinder
langchain
langchain-google-genai
langchain-community
langchain-tavily
langgraph
langchain-mcp-adapters
python-multipart
pypdf
yfinance
mcp
# Add any others if needed
